backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic>=2.6.1
motor>=3.6.0
python-dotenv==1.0.1
google-genai>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
PyPDF2>=3.0.0
python-pptx>=0.6.23
scikit-learn>=1.4.0
numpy>=1.26.0
